backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.2.0
openpyxl==3.1.2
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
python-dotenv==1.0.1
httpx==0.26.0
python-multipart==0.0.6
pytest==8.0.0
pytest-asyncio==0.23.3
